+++ types/gen/go/transport/v1/transport.pb.validate.go
@@ -2800,6 +2800,39 @@ func (m *Response_Data_Data) validate(all bool) error {
 
 	}
 
+	if m.Kinesis != nil {
+
+		if all {
+			switch v := interface{}(m.GetKinesis()).(type) {
+			case interface{ ValidateAll() error }:
+				if err := v.ValidateAll(); err != nil {
+					errors = append(errors, Response_Data_DataValidationError{
+						field:  "Kinesis",
+						reason: "embedded message failed validation",
+						cause:  err,
+					})
+				}
+			case interface{ Validate() error }:
+				if err := v.Validate(); err != nil {
+					errors = append(errors, Response_Data_DataValidationError{
+						field:  "Kinesis",
+						reason: "embedded message failed validation",
+						cause:  err,
+					})
+				}
+			}
+		} else if v, ok := interface{}(m.GetKinesis()).(interface{ Validate() error }); ok {
+			if err := v.Validate(); err != nil {
+				return Response_Data_DataValidationError{
+					field:  "Kinesis",
+					reason: "embedded message failed validation",
+					cause:  err,
+				}
+			}
+		}
+
+	}
+
 	if m.Cosmosdb != nil {
 
 		if all {