+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - 403 status codes from the server now return as 403 instead of 500
 - Added Superblocks partner ID to the Databricks integration
 - Fix scheduled job execution for signed jobs (when signature verification is enabled)
+- Now return Kinesis Plugin Metadata
 
 ## v1.15.1